Add functionality to activeIds comparison
I'm not entirely happy with the typing here. I'd like it to accept the
two specific types as being required symetrically. ie the two params
must be either ActiveDegreeIds or ActivePitchIds. It doesn't like that
though, and seeing as how I know that the pitch and degreeids will never
have the same text this is safe, for now..