Add reducers for remaining covariant global props
And because they are both of the PitchIds type they can share a
comparison function which means that it needs to go up in to the utils
folder continuing the messing up of that folder (to be decluttered
later).

I've wondered in a TODO comment in this commit whether there comparison
functions should be added to the harpparts project.