Add the covariant reducers to the newharpstrata reducer
It wasn't strictly necessary to add the new props to the mock objects in
this commit because the reducers are actually ok to recover undefined
from the global object.

However, if the reducers get more complex then we could end up with
those tests failing and the reason looking quite mysterious. So might as
well get it done.