Take harpstrata setter out of callback
I have to review why so many of these functions are in callbacks
actually. I thought it was required to prevent the menu rerendering and
the list being scrolled back to the top but that doesn't seem to be
happening here.

The point of this branch is to explore the possibility or reducing the
number of state updates by using hooks for the interaction callbacks
instead of reducers so I'm going to plough ahead and come back to this
consideration later.