Demo that pseudo-callback state updates can isolate
By which I mean that if you go to the covariant menu, make sure that the
root pitch is the static variable and modify the harpstrata via the
other two variables then you will that the NewStateLogger is not
rerendered, because only the `activeHarpStrata` state was modified, and
not the `rootPitchId`. If you change the `rootPitchId` then you will see
that the `NewStateLogger` is rerendered due to the callback effect in
the `NewHarpStrataStateCallback` component.

However, I have just noticed that the rerendering concern I had in the
previous commit is actually valid. I was mistaken about what would be
rerendered. It's not the list, it's the stacked menus. So I do need to
go back and see how the stacked menus can avoid reverting to 0 index
each time I call that `itemTapHandler` created in the previous commit.