Replace hook with reducer again
It looks like there is no way to use the hook approach from within a
callback unless you pass in the properties which are required, and doing
that means that you need to have a hook outside the callback to collect
the state which is required within, and then you might as well not have
the callback in the first place.

I'm sure there are other ways around the problem I was facing there, but
the one I've found is to return to using reducers via dispatch, but to
specifically dispatch to the one state property which is being updated.
That way we can prevent a full global state update but we don't need to
pass the property *being* updated in to the callback.

We do still need to pass the `activeDisplayMode` in to the callback
though, since not doing so will mean the callback doesn't know when the
`activeDisplayMode` has been udpated and can't consider it in it's
action.

We can use this pattern in the other menu's too, but first I would like
to build out the callback components which will update the derived
global state properties when the harpstrata is updated to check that the
rest of this concept works.