Demo that all derived state can be set from one component
If you use the covariants menu now you will see that only the relevant
state item is updated on each interaction. As you use different stacks
in the menu only 2 of the covariant states are updated at a time.

Moreover, you can see that we only get a log from the NewStateLogger
when the root pitch is updated which doesn't happen on the lowest stack
in that menu.

The logging is a little wrong on the callback hooks actually, because
what we are doing is printing out the previous state's value rather than
the one which is about to be set. But that's ok, it still proves the
point.