Rough add remaning global state deriver hooks
The way these are written assumed a `activeHarpStrata` origin of all
updates. I would like to see what happens if we make the `useEffect`
hooks actually only fire on the modification of the derivative source
variables, and to make the deriver functions actually able to utilise
other derived properties. The prime example is of the
`viewableDegreeMatrix` property which itself is derived in part from the
`columnBounds`. In this iteration I've decided to have the
`viewableDegreeMatrix` derive it's own `columnBounds` rather than rely
on being triggered by the update of the actual `columnBounds` global
state property because I'm concerned that this will produce extra hook
calls. I'm not convinced that those extra hook calls would be a problem
though so I'm going to test the water with that in a subsequent commit.

The other point to note in this commit is that all of this logic is
copied directly from the ..new-harpstrata.. reducers in the root utils
folder which is why such a lot of code has been possible to add in this
single commit. There was a shortage of unit tests for those reducers
because they were knocked up as a POC, and consequently there is still a
shortage of unit tests here. I will need to get round to them where
relevant, but I do think that the way the derivative and then the
comparison functions (which do actually mostly have tests) is done makes
the code fairly self documenting, if not rigorous at this point.