Replace global reducer with prop in tunings menu
This commit also sees the first small step in to using a derived
property to drive something. The `activeDegreeMatrix` is driving the
zoom menu update, which is not ideal because it means the menu stack
resets to bottom each time the degree matrix is updated (which is every
time anything in this menu changes basically).

A better solution would be to have the individual items passing that
prop like they do with the callback prop. That'll take some reworking
though.

There is also a problem with the columnBounds callback. It seems like
it's not registering a change properly. I think it might have something
to do with being modified twice. I would like to try and resolve this by
changing the callback deriver functions so that they are actually only
driven by the properties they care about, even if that is another
derived property. Time to experiment.