Move all derivative state updates in to shared useEffect
This may not be necessary, but it is the most likely setup to produce
something like simultaneous updates in all the states which are being
updated.

I am experimenting with using the `layoutFacts` inside the `useSizes`
hook, but the results are not satisying. There is a delay between
certain elements being rerendered and others being rerendered. I think
this might be to do with the fact that most components are still
actually being driven in to rerender by the `activeHarpStrata` which is
happening first, and then others are only being driven afterwards by the
`useSizes` hook.

This is an all or nothing experiment, it seems. I'll have to remove the
use of `activeHarpStrata` from everything to see whether my theory is
correct. If it is then I can consider moving back to the use of
separate `useEffect` hooks for every state prop.