Remove all hook reference to `activeHarpStrata`
This seems to be working. It isn't clear that it's an improvement yet,
but we'll just have to keep going.

The issues with unsyncronised renders has been resolved. It will be
interesting to see if it remains resolved if we split the deriving
functions back out in to their own `useEffect` hooks, but it's too early
for that yet, we still need to resolve an issue with the menu's whereby
the production validation that not hooks are being called within the
menu is flagging and throwing. Time to implement an idea I had earlier.

This commit doesn't actually completely remove the use of
`activeHarpStrata` either. It remains in use in some of the reducers in
quis mode etc but I will get to them once the main points of this
concept have been proven.