Extend the OptionItem component for extra global props
This is not the most elegant solution but it will do while the point is
being proven.

Ideally we would have a generic property which could stand in all the
situations required. Perhaps that would be unnecessarily confusing to
understand though so perhaps all that's required is some more formal
extension than what we have right now which is just code duplication.

This resolves the problem we were having though which is excellent. The
menu's now only make use of the `useGlobal` hooks inside the callback
functions so no rerender occurs when items are selected.