Address rendering wrinkle in quiz mode
Because previously the cycle ignored the possibility that a cell had
been interacted with between the answer flush happening and the
buffer bypass reset happening, there was the possibility that a cell
would render itself as active, out of sync with the activeHarpStrata
state which would look out of place.

The solution here is to add a flush step to the Ask phase of the quiz
questioning. During that phase it's impossible for the user to interact
with the cells which means that they can't buffer up any more toggles,
whcih means it's safe to do the hard reset coming in to the Listen stage
again.

I think there's probably more improvements to be made here, knowing now
what I know about the relative benefits of dispatchers and `use..`
hooks. It is out of scope for this feature though which is targeted
soley at removing the use of the `activeHarpStrata` `useGlobal` hook to
prevent unnecessary renders. I think that has been achieved here now.