Replace hook with inline defined dispatches
The hook was obscurring the functionality unnecessarily and since there
are so many inline defined reducers defined now anyway, it is more
consistent to move it out like this.

The key feature is that we can see which additional properties are being
hooked in to for the dispatcher to work which was obscured when using
the hook directly. That could be considered an advantage too, but
consistency is really the most important thing here.