Add viewableInteractionMatrix derived global state
This is necessary to do away with a couple of hooks we've got in use
currently which were brought in before this derived state plan was fully
realised.

This commit also sees the new state used as the source of the
layoutFacts global state because it is less volatile than the
activeDegreeMatrix state and will produce fewer runs of that deriver
function.