Remove another unused hook
In this commit you can actually see that I had one of the tests setup
incorrectly. The global `layoutFacts` property wasn't actually being
used in this test because the component was previously using the hook
which was getting it's information from the matrices rather than the
`layoutFacts` global state property.