Add source property for columnBounds
This is necessary so that the update following a columnBounds change is
all at once. If we set the `columnBounds` directly then the updates to
the viewable matrices and the layoutFacts happens shortly after which
creates an unpleasant moment in the rendering.

The naming is still a little in flux. Probably I will want to use the
word "source" elsewhere too, and I will probably want to use the setter
naming convention I've started in this commit elsewhere too. I'll trial
that in the next commit in fact.