Remove addressed TODOMANRENDER comments
Now that we have exploded out the various global states we have actually
achieved all of these.

The only exception is the one on `use-zoom-items` which calls for a
dedicated zoomid state which is overkill. The labels can just be
rerendered each time the columnBounds are scrolled. I don't think that
is such a burden. And if it is then we can deal with it when we get that
feature added.