Remove unused file & rehome component util
The isMatchPozitionIds util isn't really required and probably belongs
being defined in the harpparts project, but let's just keep it alive
here for now until we've got time to do that.

I've reduced the specificity of the equality check too since it's not
important that the onjects are identical, it's just important that they
match, even if right now the two are the same thing.