Return the initial load scale flash
The approach needs to be a bit different to before.

It's not possible to set the `bufferedActivityToggles` property and wait
for the toggle flusher to take effect because immediately upon setting
the `activeHarpStrata` the callback function associated with that will
empty the `bufferedActivityToggles`. Now we just have to set it directly
and have the notification display hook start from a dead start with
`undefined` as it's initial value to compare against.