Repurpose util for precision
The pattern of functionality here is that we want to reduce from the
most relevant source. In this case that is one of the full matrices, not
the harpstrata.

I've also taken the opportunity to keep some of the variables naming in
sync. This "full" prefix is catching on and there's no point adding it
to the interaction matrix but not the degree and pitch matrix within the
same file.