Combine source global state callback in single component
For those who don't want to read the diff; the purpose of the two
components being replaced was very similar. One responded to updates to
`activeHarpStrata` (transitioning to the name `sourceHarpStrata`) by
updating the dependent global properties at once.

The other did the same but for the `souceColumnBounds` property.

Having them in one place makes a lot of sense not only because of the
similarity of function, but also because we can possibly bring some of
the dependant utils in to this one folder rather than having them up in
the src utils folder.