Remove unused testing component
I haven't used this to any more effect than proving the fact that having
a *source* global state prop which is never rendered on, with some
derivative state props which can all be updated *at once* from it is an
effective optimisation for one or two properties.

Perhaps I should do a more detailed check of how well I have implemented
that idea but the evidence in the UI experience directly is convincing.
I won't take the time for now. If further optimisation is required then
I should check the quality of this framework first.