Make dedicated "menu aware" component
This means that the `useMenu` state hook is no longer called in the
HarpGuruPage component meaning that that component doesn't need to
rerender every time the menu state changes.

However, there is still a significant delay when hitting the menu
buttons. I think this is because the menu's themselves have reanimated
variables defined in them which are heavy to rerender, and each of the
menus is currently being rerendered each time on is interacted with.

Possibly some memoisation would be helpful here.