Remove unused local, accidentally commited index
There is a local folder for a candidate component called
option-item-with-fullmatrix which has been sitting here for a while. I
didn't realise I accidentally committed the index file. That will mean
that the last few commits I've made would actually fail remotely where
the rest of this component is not defined.