Add unit tests to reducer and enhance functionality
It's not strictly necessary for this reducer to perform this check the
way it's working currently. However it does no harm and it is arguably
just a feature of reducers which is wise to have incase they are used in
a dispatch in the future.

Actually a further change should be made to build that future because in
those situations the first parameter is the same as the return type when
a dispatch is being used.