Make reducer amenable to dispatchers for futureproofing
This actually brings the intention of this reducer in line with ones
which are actually being used in dispatchers at the moment.

I should update the naming convention of those other reducers to match
these for consistency now that they are equivalent.

I also need to reproduce this enhancement for all the other reducers in
this set of functionality.