Add unit tests for very specific case
It is important (conceptually) to favour returning the previous property
as identify if a match is found. There was an edge case in this logic
before this change whereby that would be done *except* if the
columnBounds were 'FIT' in which case the `fullMatrix` would always be
returned. Depending on usage this could result in unnecessary renders.