Replace global state setter util with dispatch
The setter is effectively just doing what the dispatch is already
designed to do. This is much more efficient and we will see in the
sourceHarpStrata callback hook next that we can reduce the amount of
code we need dramatically by doing this.

One thing we lose with this strategy is that we no longer have access to
the property returned from the reducer when it's called inside the
dispatcher. Previously we could ensure we recieved the very same object
back from the setter as was just being set. Now we need to run the
reducer outside the dispatch context as well as inside of it. This is a
little less efficient, but the readability benefits outweigh this cost.