Add setter to workaround dispatch shortfall
After raising this issue
(https://github.com/CharlesStover/reactn/issues/215#issuecomment-940226958)
I'm happy to add this workaround setter which will be used instead of
the existing dispatches in this components hooks.

I will keep the reducers as they are, leaving the logic which determines
whether to return the previous or next property *in* the reducers so
that they would still be applicable to dispatch as would be the
"proper" design.