Improve test data to avoid false positives
It turns out that `reduce-harpstrata-to-rootpitchid` was implemented
incorrectly, but it wasn't picked up by the test because harpKeyId and
rootPitchId are the same in first position. Updating the position of
this base test data in to second position makes this kind of false
positive less likely. Several test files have had to be updated along
with this change.