Remove zoom menu (temporarily)
The zoom menu is not part of a complete feature until we also have a
scrolling widget. Otherwise when users zoom they will have no way to
access the rest of the harp.

This is necessary because it is a good moment to do a release of the
pentaharp tuning and this is the only part of the current UI which is
incomplete.