Make zoom slide invisible if has no work to do
If the entire harp is visible then this component needn't be here.

There will be some work to do in the sizing hook to make sure it is also
aware of the visibility / non-visibility of this component so that it
can scale the rest of the harp face accordingly.

Before that I still need to prove that this is a functional component.