Improve styling of zoom slider
Just cleaning it up a bit to get a better feel for how it could fit
visually.

I'm still not convinced. Having a filled bar rather than a moving slide
doesn't feel like the right tool to convey what's happening. I've tried
to add a renderIndicator property to add a bar, but it's not worked.