Widen slide indicator (and break slider)
The intention of this commit was to make the slider proportionally
represent the amount of the total harp that we can currently see. In
doing so the pan gesture handling is no longer reachable.

I think I need to try and implement what I'm looking for without these
prefabricated components now. They've helped me understand what I'm
aiming for, but now I need to make it myself.