Add selected hole label to slider
This breaks the animation at the moment because storing the
`liveStartHole` state in the parent produces a rerender on change which
need only happen in the child really.

The effectis that when the rerender happens the animation returns to the
slide origin for a split second before continuing just fine.