Replace double ref requirement with single one
It's obviously cleaner to maintain only one ref in to a single labelling
component which spans the whole slider than to maintain two.

Having the flex styling contained in the slider rather than the labels
means that this same component can be added to a horizontal slider (if
we want one) without any changes required.