Homogeonise variables by matchin local form to global
It made sense to only store the start index, except that it was making
the variable space noisy. This simplification makes it easier to see
when we're done with the global and dealing with the local state.

The reason we need both is because animating based on the global state
producing janky behaviour since so many components are dependent on the
global columnBounds. But we need the global to know when the situation
has changed at a global level and we need to respond at a local one.