Add dedicated legendWidth to size scheme
This is relevant for the feature being designed here, so that updates
tot he legend width are handled gracfully in the repositioning of the
slider.

I can't account for the very slight change in the resulting sizing.
I did take a little liberty in this commit and shuffle some of the
size calculations around to honor the presence of the new legendWidth
property.. I'm in a rush you see, as I keep mentioning in these commits.

The size changes are insignificant and are probably just the result
of some floating point inaccuracy resulting from the reordering of
the sums slightly.

It's not the result of the inclusion of the zoomSlideWidth in the static
sizes. Although now that I think about it, I don't want that to be there
because it's a conditional size and will mean that the static sizes
change when there is a slider.

I'm making this commit message as an amendment so I'll have to rectify
this in the next commit.