Remove zoom slide from static size calculations
It's the only size which is not statically defined in it's relative form
so including it in the static size calculations means that they are not
static anymore.

Really I should just be calculating the static sizes as a function of
the screen width and then I wouldn't have these hazards.