Stash the partially complete refactor
I have needed to commit these partial refactoring because there's an
awful lot that's been done. It still isn't ready for a proper commit
because ... it's just not all shifted as I need it to yet. But I have to
commit and push something or risk losing a whole days worth of partial
refactoring.

If this commit ever finds it's way on to the master branch then it's
because something horrible happened and I was extremely glad that I made
this commit.

I have had to disable husky to be able to make this commit, but I
haven't committed that change.