Lay foundtions for double reedarray tunings (chromatics)
The number of changes in this commit are absolutely necessary. Allowing
a tuning which has two sets of reed arrays is a change which effects the
foundations of the harp parts definitions all the way up to the
rendering logic.

The use of a tuple for all tuning definitions keeps the form homogeonous
for across exisitng harps and future chromatics while creating space for
the chromatics too. The consistency is really important for this stage
because it makes conceiving of the concequences more managable. It might
not be the ideal design once things are working, but after a lot of
thought it is the right design for these early stages.

This commit does leave some things broken. That is probably a sign of
missing tests. Some of those tests are not possible (or simple enough)
to create due to `useEffect` hooks or animation values. Some though are
just missing so reparining the functionality will be a good opportunity
to fill in some gaps in the tests.

I do not want to postpone making this commit until I've fixed those
other issues because there is already plenty here, and passing the
existing tests is a decent & logical milestone.