Prepare to improve typesafety of double harpface
The previous foundation laying proved quite a lot, but it also proved
that we could use a bit more type safety. Defining our different types
of harpfaces as a tuple left a lot of room for accidentally picking the
wrong dimension of an array since the contents of those tuples were also
arrays, hense picking on array level too shallow would not produce a
type error.

Using this new structure should avoid that and mean that we don't
actually need to write any additional tests to fix the problems of the
previous commit. Writing tests will be much more work and be less
guarantee of success and give less immediate feedback in future work.