Improve typesafety of multiple harpfaces
Using an object definition like this rather than a tuple means that we
have greater type safety because we aren't wrapping a multidimensional
array within another array. We are much more explicit about what we are
interacting with.

This commit isn't perfect. Amongst other things it raises the question
of how we should be arranging the layout facts object equivaluent to the
harpfacematrices. I am also still not settled on 'harpface1' &
'harpface2' as the property names for the HarpFaceMatrices object.