Add a chromatic tuning and ability to render
This works quite nicely.

I have hardwired the octave groupings to be based on the top row (where
the harp slide is out) because I think having a staggard octave grouping
between the top and bottom would be confusing. Moreover, if we remove
the bends from being displayed then the logic as it stands won't
actually draw any octave boundaries in. Actually, this is going to be a
problem for certain positions if we get rid of the bends, regardless..
I'm going to have to change the octave column grouping logic so that it
can potentially use both and just return a mapping for whichever
harpface has some root notes in it.