Rename for differentiation
It's maybe slightly clearer with this naming that we will sometimes be
dealing with a plural of harpfaces. I considered ..HarpFaces.. but then
that doesn't play well with the already existing HarpFaceMatrices.
Perhaps HarpFacesMatrix, would seem better but when you think about the
actual global property names which will relate to that type
(eg fullDegreeMatrices) it seems like this is a better compromise.