Add a harpfactfacts mapper function
Since harpfacefacts are going to be used more widely and since we seem
to have a lot of occasions where we're going something quite similar to
what this util does, but inline, in a number of HarpFaceMatrices
locations at the moment, it seems wise to centralise the logic.

The next commit will see us take a turn in this direction with
reedarrays so we'll get to put it to the test soon.