Utilise alternate global prop
In a hopeless attempt to simplify the hooks job. The
viewableInteractionMatrix property is already used in the layoutFacts
global property derivation so I thought perhaps this was more direct.
That logic is definitely flawed though since all global properties are
updated at the same time from the source global props. I'm leaving this
commit in place though because I've just done some comparative speed
tests which I've recorded the results for in
https://github.com/js-jslog/harpguru/issues/134, and I've raised
another issue to deal with the comment which is removed here in
https://github.com/js-jslog/harpguru/issues/133