Use typeguard to replace adhoc type checking
This wouldn't have strictly fitted in to the feature change which it's
being performed in. However, the feature change (adding chromatics) lead
to a review of the function (as you can see from the TODO above the
changed lines in this commit), and this became a relevant change as a
result.