Rename function to prepare for new global prop
There is a serious flaw with the `getIfZoomedColumnBounds` function. The
layout facts that it uses at the moment are only the viewable matrices.
That means that comparing them to the columnBounds will not tell us
anything about whether the harpface is zoomed or not.

I plan to identify add layout facts for both the viewable and the full
matrices.