Add a proper equality checker to the size reducers
The plan originally was to rely on the fact that all of the sizes are
intrinsically related so that if one has changed then they will all have
changed. That is not actually true though. It's possible for the
zoomSlideWidth to change from a value to 0 without the seed sizes being
a factor. It's probably just safer to check everything, even though at
the moment we could get away with just testing one of the sizes and the
zoomSlideWidth together.