Rename reducer to better suite situation
There is no layoutFacts involved in this so there's no need for that in
the name.

A more time consuming but slightly better change might be to remove this
property from the global state since it's easily and non-expensively
derived as a util. However, there's something to be said for consistency
when it comes to recovering all size information from a similar place
whether it's static or dynamic so I'm undecided here still.