Remove 3rd page from page collection
I'm not sure anyone needs the third page. It was useful for the
demonstration of the 3 positions of the 12 bar blues, but actually I
think most people (if they even use multiple pages) want to compare one
page directly to another and switching back and forth is definitely the
most effective way to do that.