Clarify chromatic tunings relationships
The new file `chromatic-tunings.ts` was supposed to give a way to derive
all of the chromatic tunings from a single "mother-tuning". The concept
was fine but for some reason produced a very difficult issue at
compile-time. The problem seemed to occur around the running of
`mapHarpFaceFacts` and lead to the file in which that function was being
run as being `undefined` during import. I may try again or to just use
that logic as part of a test file to validate the manually written
tunings if that is more acceptable to the compiler.

I have also duplicated the `slice-matrix` package from `harpguru-core`.
This probably means it should be made in to it's own package up a level
higher and give an package.json etc, but I'll save that for the tidying
stages.