Add reanimated dependencies in app root
I don't understand why this is necessary when using eas.
It wasn't necessary when running expo go, but without
this all of the binaries produced error on load.

The reason it shouldn't be necessary is because
reanimated is only used in the harpguru-core package
which has it's own dependency on these packages.

Nevertheless, it is necessary and for now they'll have
to be here. I may raise a bug ticket / question about
this