Set to run Hermes engine
The app opens a few seconds quicker with this.
I see no downside.

I'm adding this in this branch even though it's
off topic because it was one of the thingns
that I tinkered with while trying to get
reanimated working. It's completely unrelated
but I'm indulging because of the urgency to get
across the finish line and how little patience
I have for managing further merge requests at
this point.