Prepare for next tag
It has not been necessary to increase the android version number because
there was no build made at the previous tag. The reason being that this
new tag comes after a fix to enable a build to even be made.

The iOS build number was increased in order to stay in sync with the
monorepo's versioning, even though for iOS there also was no release at
the previous tag.