Create user-manual.md
The target of this document is the end user rather than the developer. That means it's a little out of place in this folder, but this is the most time efficient way of making the instructions available to the users for the moment.

In other words, this is not ideal, but it's benefits outweigh the cost.