Install a new expo project
I've stripped down the pacakge.json and yarnlock files so that we're
starting from scratch. I'll do an install in the next commit and see if
it still works.

I've tested that the cool-app project starts and serves with
`EXPO_USE_METRO_WORKSPACE_ROOT=1 npx expo start --tunnel`.

I've had to remove husky for now so that we can build up slowly and the
lerna.json because I don't think I'll need it in the end.