Add a pre-push hook
Since I've let go of lerna I'm not too sure how to do lint-staging in a
similar way to before in this yarn workspace.

I'm going to skip that for now since the release of the new Android SDK
is so urgent. I'll not be able to actually push any changes with any
issues present in the project with this hook so I'm protected, if not
with the same responsiveness as before.