Upgrade failing dependency
Narrowed the search down to this dependency, updated it and found it
worked. Simple as that.

I'm afraid to say that I actually don't know that this dependency is
required anymore. I removed it and things seemed to still work. But
unfortunately I'm at a stage with this project at the moment where I
don't have the time to reassure myself further than running the test
suite. As much as this is a capitulation to the project fully being in
life support mode I do not want to take the (apparently unnecessary)
risk.

I can point to the original introduction of the original inclusion of
the hook and identify that the reason given here would seem to still
apply: ebadd6c26dd2dce

Of course there is also risk with introducing a major version change
too. Perhaps it's almost equivalent. But I do think it's at least
slightly lower. The basic functionality of the hook as I need it (to
have the hook triggered when the window size changes) is unlikely to
have changed.